Refactor namespaces to FeatureManagement
