+++ source/ProcessManager.Core/Application/FeatureManagement/FeatureFlag.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Core.Infrastructure.FeatureFlags;
+using Energinet.DataHub.ProcessManager.Core.Infrastructure.FeatureManagement;
 
-namespace Energinet.DataHub.ProcessManager.Core.Application.FeatureFlags;
+namespace Energinet.DataHub.ProcessManager.Core.Application.FeatureManagement;
 
 /// <summary>
 /// The current feature flags in the Process Manager. If using <see cref="MicrosoftFeatureFlagManager"/>
